src/requirements.txt
-i https://pypi.org/simple
alembic==1.9.2
anyio==3.6.2 ; python_full_version >= '3.6.2'
async-timeout==4.0.2 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
click==8.1.3 ; python_version >= '3.7'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
email-validator==1.3.1
fastapi[all]==0.89.1
greenlet==2.0.2 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httptools==0.5.0
httpx==0.23.3
idna==3.4 ; python_version >= '3.5'
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
orjson==3.8.5
psycopg2-binary==2.9.5
pydantic==1.10.4 ; python_version >= '3.7'
python-dotenv==0.21.1
python-multipart==0.0.5
pyyaml==6.0
redis==4.4.2
rfc3986[idna2008]==1.5.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.41 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy2-stubs==0.0.2a32 ; python_version >= '3.6'
sqlmodel==0.0.8
starlette==0.22.0 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
ujson==5.7.0
uvicorn[standard]==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
xlsxwriter==3.0.8
